HTTP/2: fixed handling of output HEADERS frames.
The HEADERS frame is always represented by more than one buffer since
bff32110ecca, but the handling code hasn't been adjusted.

Only the first buffer of HEADERS frame was checked and if it had been
sent while others had not, the rest of the frame was dropped, resulting
in broken connection.

Before bff32110ecca, the problem could only be seen in case of HEADERS
frame with CONTINUATION.

--HG--
branch : nginx