SSL: fixed possible segfault on renegotiation (ticket #845).
Skip SSL_CTX_set_tlsext_servername_callback in case of renegotiation.
Do nothing in SNI callback as in this case it will be supplied with
request in c->data which isn't expected and doesn't work this way.

This was broken by 8b907c73cf81 (1.9.6) with OpenSSL master branch and LibreSSL.

--HG--
branch : nginx