Cache: prefix-based temporary files.
On Linux, the rename syscall can be slow due to a global file system lock,
acquired for the entire rename operation, unless both old and new files are
in the same directory.  To address this temporary files are now created
in the same directory as the expected resulting cache file when using the
"use_temp_path=off" parameter.

This change mostly reverts cef756b5d206 and 47be8a01e6a4, restoring the
behaviour as of 5fec36df9bad (with minor changes).

--HG--
branch : nginx